+++ bases/criterium/src/criterium/sampled_fn.clj
@@ -11,6 +11,7 @@
    [criterium.measured :as measured]
    [criterium.metric :as metric]
    [criterium.sampler :as sampler]
+   [criterium.util.forms :as forms]
    [criterium.util.invariant :refer [have]]
    [criterium.util.t-digest :as t-digest])
   (:import
@@ -22,7 +23,7 @@
    (fn measured-f [args ^long eval-count]
      (have (partial = 1) eval-count)
      (let [start  (jvm/timestamp)
-           res    (apply original-fn args)
+           res    (forms/unrolled-invoke original-fn args)
            finish (jvm/timestamp)]
        [(unchecked-subtract finish start) res]))))