+++ components/utils/test/criterium/utils/forms_test.clj
@@ -1,7 +1,28 @@
 (ns criterium.utils.forms-test
   (:require
    [clojure.test :refer [deftest is testing]]
-   [criterium.utils.interface :refer [cond*]]))
+   [criterium.utils.interface :refer [cond* unrolled-invoke]]))
+
+(deftest unrolled-invoke-test
+  (testing "unrolled-invoke applies f to all elements of the args collection"
+    (testing "zero args"
+      (is (= :ok (unrolled-invoke (fn [] :ok) []))))
+    (testing "one arg"
+      (is (= 1 (unrolled-invoke inc [0]))))
+    (testing "several args"
+      (is (= 6 (unrolled-invoke + [1 2 3])))
+      (is (= 24 (unrolled-invoke * [2 3 4]))))
+    (testing "arity at the unrolled boundary (20 args)"
+      (is (= 20 (unrolled-invoke + (vec (repeat 20 1))))))
+    (testing "arity beyond the unrolled boundary falls back to apply"
+      (is (= 25 (unrolled-invoke + (vec (repeat 25 1))))))
+    (testing "f and args expressions are each evaluated once"
+      (let [f-calls    (atom 0)
+            args-calls (atom 0)]
+        (is (= 3 (unrolled-invoke (do (swap! f-calls inc) +)
+                                  (do (swap! args-calls inc) [1 2]))))
+        (is (= 1 @f-calls))
+        (is (= 1 @args-calls))))))
 
 ;; Tests for the cond* macro from criterium.utils.interface.
 ;; cond* extends standard cond with :let bindings visible to subsequent clauses.